Django-backend/requirements.txt
Django==2.2
pandas==0.24.2
django_rest_framework_mongoengine==3.3.1
mongoengine==0.17.0
djangorestframework==3.9.4
